MT-22401: take params objects for campaign list and stats
Decisions:
- Accept EmailCampaignListParams/EmailCampaignStatsParams instead of loose kwargs
  so new query params extend the params class, matching StatsApi.get
- Export both params classes; they are public call surface now, as StatsFilterParams is
- Move Pagination to models/common.py; the payload is generic, not campaign-specific
